+++ mappers/Mapper009.h
@@ -12,15 +12,15 @@ public:
 	std::string name() const override;
 
 public:
-	void write_a(uint32_t address, uint8_t value) override;
-	void write_b(uint32_t address, uint8_t value) override;
-	void write_c(uint32_t address, uint8_t value) override;
-	void write_d(uint32_t address, uint8_t value) override;
-	void write_e(uint32_t address, uint8_t value) override;
-	void write_f(uint32_t address, uint8_t value) override;
+	void write_a(uint_least16_t address, uint8_t value) override;
+	void write_b(uint_least16_t address, uint8_t value) override;
+	void write_c(uint_least16_t address, uint8_t value) override;
+	void write_d(uint_least16_t address, uint8_t value) override;
+	void write_e(uint_least16_t address, uint8_t value) override;
+	void write_f(uint_least16_t address, uint8_t value) override;
 
 public:
-	uint8_t read_vram(uint32_t address) override;
+	uint8_t read_vram(uint_least16_t address) override;
 
 private:
 	uint8_t chr_ram_[0x2000] = {};
